+++ src/token.rs
@@ -3,7 +3,7 @@
 pub enum Token<'a> {
   // Variants that take an argument
   Identifier(&'a str),
-  Value(u32),
+  Value(u64),
 
   // Keywords: input, output, persistent, transient, constant, snippet, and, or, not, bit
   Input,
